Fixed Android unit tests
